- revise some user interface stuff to improve clarity and remove some superfluous tags.  Modify input files appropriately - rework the FieldAdaptor.h information to fix the field typing.  This is the critical change for this checkin.  This also goes along with some recent changes in SpatialOps. - quiet some warnings (TaskInterface.cc, TimeStepper.cc) that were showing up on my linux box. - add a new input file to test creation of equations on staggered cells.

+++ src/CCA/Components/Wasatch/TimeStepper.cc
@@ -205,13 +205,13 @@ namespace Wasatch{
   {
     //__________________
     // loop over patches
-    for( size_t ip=0; ip<patches->size(); ++ip ){
+    for( int ip=0; ip<patches->size(); ++ip ){
 
       const Uintah::Patch* const patch = patches->get(ip);
 
       //____________________
       // loop over materials
-      for( size_t im=0; im<materials->size(); ++im ){
+      for( int im=0; im<materials->size(); ++im ){
         
         const int material = materials->get(im);
